Add test cases for 'Recorder' mechanism.

+++ src/test/java/com/github/dakusui/jcunit/framework/tests/bugfixes/geophile/GeophileRecorderTest.java
@@ -1,35 +1,38 @@
 package com.github.dakusui.jcunit.framework.tests.bugfixes.geophile;
 
+import com.github.dakusui.jcunit.core.JCUnit;
 import com.github.dakusui.jcunit.core.Utils;
 import com.github.dakusui.jcunit.core.rules.JCUnitRecorder;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.Rule;
 import org.junit.Test;
+import org.junit.runner.RunWith;
 
 import java.io.File;
 
+import static org.hamcrest.CoreMatchers.is;
+import static org.hamcrest.MatcherAssert.assertThat;
 import static org.junit.Assert.assertTrue;
 
+@RunWith(JCUnit.class)
 public class GeophileRecorderTest extends GeophileTestBase {
-  private static final String RECORDER_BASE = "src/test/resources";
+  private static final String RECORDER_BASE = null;
   @Rule
-  public JCUnitRecorder recorder = new JCUnitRecorder(RECORDER_BASE);
+  public JCUnitRecorder recorder = new JCUnitRecorder();
 
   @BeforeClass
   public static void beforeClass() {
     File baseDir =Utils.baseDirFor(RECORDER_BASE, GeophileRecorderTest.class);
-    Utils.deleteRecursive(baseDir);
+    if (baseDir.exists()) {
+      Utils.deleteRecursive(baseDir);
+    }
     assertTrue(!baseDir.exists());
   }
 
   @Test
   public void test() {
-  }
-
-  @AfterClass
-  public static void afterClass() {
     File baseDir =Utils.baseDirFor(RECORDER_BASE, GeophileRecorderTest.class);
-    assertTrue(baseDir.exists());
+    assertThat(baseDir.list().length, is(this.recorder.getId() + 1));
   }
 }
